Move dispatch stuff into `CubedSpheres` module
