+++ validation/cubed_sphere_tracer_advection/cubed_sphere_tracer_advection.jl
@@ -11,72 +11,14 @@ using Oceananigans.Coriolis
 using Oceananigans.Models.HydrostaticFreeSurfaceModels
 using Oceananigans.TurbulenceClosures
 
-#####
-##### Gotta dispatch on some stuff after defining Oceananigans.CubedSpheres
-#####
+using Oceananigans.Diagnostics: accurate_cell_advection_timescale
 
-using Oceananigans.Models.HydrostaticFreeSurfaceModels: ExplicitFreeSurface
-
-import Oceananigans.CubedSpheres: maybe_replace_with_face
-
-maybe_replace_with_face(free_surface::ExplicitFreeSurface, cubed_sphere_grid, face_number) =
-  ExplicitFreeSurface(free_surface.η.faces[face_number], free_surface.gravitational_acceleration)
-
-import Oceananigans.Diagnostics: accurate_cell_advection_timescale
-
-function accurate_cell_advection_timescale(grid::ConformalCubedSphereGrid, velocities)
-
-    min_timescale_on_faces = []
-
-    for (face_number, grid_face) in enumerate(grid.faces)
-        velocities_face = maybe_replace_with_face(velocities, grid, face_number)
-        min_timescale_on_face = accurate_cell_advection_timescale(grid_face, velocities_face)
-        push!(min_timescale_on_faces, min_timescale_on_face)
-    end
-
-    return minimum(min_timescale_on_faces)
-end
-
-import Oceananigans.OutputWriters: fetch_output
-
-fetch_output(field::ConformalCubedSphereField, model, field_slicer) =
-    Tuple(fetch_output(field_face, model, field_slicer) for field_face in field.faces)
-
-import Base: minimum, maximum
-
-minimum(field::ConformalCubedSphereField; dims=:) = minimum(minimum(field_face; dims) for field_face in field.faces)
-maximum(field::ConformalCubedSphereField; dims=:) = maximum(maximum(field_face; dims) for field_face in field.faces)
-
-minimum(f, field::ConformalCubedSphereField; dims=:) = minimum(minimum(f, field_face; dims) for field_face in field.faces)
-maximum(f, field::ConformalCubedSphereField; dims=:) = maximum(maximum(f, field_face; dims) for field_face in field.faces)
-
-using Oceananigans.CubedSpheres: ConformalCubedSphereFunctionField
-
-import Oceananigans.BoundaryConditions: fill_halo_regions!
-import Oceananigans.CubedSpheres: fill_horizontal_velocity_halos!
-
-fill_halo_regions!(::ConformalCubedSphereFunctionField, args...) = nothing
-
-# Forget about filling velocity halos when `velocities = PrescribedVelocityFields`
-fill_horizontal_velocity_halos!(u::ConformalCubedSphereFunctionField, v, arch) = nothing
-fill_horizontal_velocity_halos!(u, v::ConformalCubedSphereFunctionField, arch) = nothing
-fill_horizontal_velocity_halos!(u::ConformalCubedSphereFunctionField, v::ConformalCubedSphereFunctionField, arch) = nothing
-
-import Oceananigans.CubedSpheres: maybe_replace_with_face
-
-maybe_replace_with_face(velocities::PrescribedVelocityFields, cubed_sphere_grid, face_number) =
-    PrescribedVelocityFields(velocities.u.faces[face_number], velocities.v.faces[face_number], velocities.w.faces[face_number], velocities.parameters)
-
-import Oceananigans.CubedSpheres: fill_horizontal_velocity_halos!
-
-# We will use prescribed velocity `Field`s
-fill_horizontal_velocity_halos!(u, v, arch) = nothing
+Logging.global_logger(OceananigansLogger())
 
 #####
 ##### state checker for debugging
 #####
 
-# Takes forever to compile with Julia 1.6...
 function state_checker(model)
     fields = model.tracers
 
@@ -122,8 +64,6 @@ end
 
 ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"
 
-Logging.global_logger(OceananigansLogger())
-
 dd = DataDep("cubed_sphere_32_grid",
     "Conformal cubed sphere grid with 32×32 grid points on each face",
     "https://github.com/CliMA/OceananigansArtifacts.jl/raw/main/cubed_sphere_grids/cubed_sphere_32_grid.jld2",
@@ -134,19 +74,25 @@ DataDeps.register(dd)
 
 cs32_filepath = datadep"cubed_sphere_32_grid/cubed_sphere_32_grid.jld2"
 
-central_longitude = (0, 90, 0, 180, -90, 0)
-central_latitude  = (0, 0, 90, 0, 0, -90)
+central_longitude = (0, 90,  0, 180, -90,   0)
+central_latitude  = (0,  0, 90,   0,   0, -90)
 
-function tracer_advection_over_the_poles(face_number)
+"""
+    tracer_advection_over_the_poles(; face_number, α)
 
-    H = 4kilometers
-    grid = ConformalCubedSphereGrid(cs32_filepath, Nz=1, z=(-H, 0))
+Run a tracer advection experiment that initializes a cosine bell on face `face_number`
+and advects it around the sphere over 12 days. `α` is the angle between the axis of
+solid body rotation and the polar axis (degrees).
+"""
+function tracer_advection_over_the_poles(; face_number, α)
+
+    grid = ConformalCubedSphereGrid(cs32_filepath, Nz=1, z=(-1, 0))
 
     ## Prescribed velocities and initial condition according to Williamson et al. (1992) §3.1
 
+    period = 12days  # Time to make a full rotation (s)
     R = grid.faces[1].radius  # radius of the sphere (m)
-    u₀ = 2π*R / (12days)  # advecting velocity (m/s)
-    α = 0  # angle between the axis of solid body rotation and the polar axis (degrees)
+    u₀ = 2π*R / perioid  # advecting velocity (m/s)
 
     # U(λ, φ, z) = u₀ * (cosd(φ) * cosd(α) + sind(φ) * cosd(λ) * sind(α))
     # V(λ, φ, z) = - u₀ * sind(λ) * sind(α)
@@ -164,9 +110,6 @@ function tracer_advection_over_the_poles(face_number)
         end
     end
 
-    # fill_halo_regions!(Ψᶠᶠᶜ, CPU())
-    # fill_halo_regions!(Ψᶠᶠᶜ, CPU()) # get those corners
-
     for (f, grid_face) in enumerate(grid.faces)
         for i in 1:grid_face.Nx+1, j in 1:grid_face.Ny
             Uᶠᶜᶜ.faces[f][i, j, 1] = (Ψᶠᶠᶜ.faces[f][i, j, 1] - Ψᶠᶠᶜ.faces[f][i, j+1, 1]) / grid.faces[f].Δyᶠᶜᵃ[i, j]
@@ -175,9 +118,6 @@ function tracer_advection_over_the_poles(face_number)
             Vᶜᶠᶜ.faces[f][i, j, 1] = (Ψᶠᶠᶜ.faces[f][i+1, j, 1] - Ψᶠᶠᶜ.faces[f][i, j, 1]) / grid.faces[f].Δxᶜᶠᵃ[i, j]
         end
     end
-
-    # fill_horizontal_velocity_halos!(Uᶠᶜᶜ, Vᶜᶠᶜ, CPU())
-    # fill_horizontal_velocity_halos!(Uᶠᶜᶜ, Vᶜᶠᶜ, CPU()) # get those corners
 
     ## Model setup
 
@@ -199,7 +139,8 @@ function tracer_advection_over_the_poles(face_number)
     φ₀ = central_longitude[face_number]
 
     # Great circle distance between (λ, φ) and the center of the cosine bell (λ₀, φ₀)
-    r(λ, φ) = R * acos(sind(φ₀) * sind(φ) + cosd(φ₀) * cosd(φ) * cosd(λ - λ₀))
+    # using the haversine formula
+    r(λ, φ) = 2R * asin(√(sind((φ - φ₀) / 2)^2 + cosd(φ) * cosd(φ₀) * sind((λ - λ₀) / 2)^2))
 
     cosine_bell(λ, φ, z) = r(λ, φ) < R ? h₀/2 * (1 + cos(π * r(λ, φ) / R)) : 0
 
@@ -233,12 +174,12 @@ function tracer_advection_over_the_poles(face_number)
     return simulation
 end
 
-for face_number in 1:6
-    tracer_advection_over_the_poles(face_number)
-end
+# for face_number in 1:6
+#     tracer_advection_over_the_poles(face_number)
+# end
 
-include("animate_on_map.jl")
+# include("animate_on_map.jl")
 
-for face_number in 1:6
-    animate_tracer_advection(face_number)
-end
+# for face_number in 1:6
+#     animate_tracer_advection(face_number)
+# end
